moved common api code out of datasets
