+++ api/dataset.go
@@ -1,46 +1,18 @@
 package api
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"time"
 
 	errs "github.com/ONSdigital/dp-dataset-api/apierrors"
 	"github.com/ONSdigital/dp-dataset-api/models"
-	"github.com/ONSdigital/dp-dataset-api/store"
 	"github.com/ONSdigital/go-ns/common"
 	"github.com/ONSdigital/go-ns/log"
 	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
 	"gopkg.in/mgo.v2/bson"
-)
-
-const (
-	datasetDocType         = "dataset"
-	editionDocType         = "edition"
-	versionDocType         = "version"
-	observationDocType     = "observation"
-	downloadServiceToken   = "X-Download-Service-Token"
-	dimensionDocType       = "dimension"
-	dimensionOptionDocType = "dimension-option"
-
-	// audit actions
-	getDatasetsAction = "getDatasets"
-	getDatasetAction  = "getDataset"
-	getEditionsAction = "getEditions"
-	getEditionAction  = "getEdition"
-	getVersionsAction = "getVersions"
-	getVersionAction  = "getVersion"
-
-	// audit results
-	actionAttempted    = "attempted"
-	actionSuccessful   = "successful"
-	actionUnsuccessful = "unsuccessful"
-
-	auditError = "error while attempting to record audit event, failing request"
 )
 
 func (api *DatasetAPI) getDatasets(w http.ResponseWriter, r *http.Request) {
@@ -1194,180 +1166,4 @@ func mapResults(results []models.DatasetUpdate) []*models.Dataset {
 		items = append(items, item.Current)
 	}
 	return items
-}
-
-func handleErrorType(docType string, err error, w http.ResponseWriter) {
-	log.Error(err, nil)
-
-	switch docType {
-	default:
-		if err == errs.ErrDatasetNotFound || err == errs.ErrEditionNotFound || err == errs.ErrVersionNotFound || err == errs.ErrDimensionNodeNotFound || err == errs.ErrInstanceNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-		}
-	case "edition":
-		if err == errs.ErrDatasetNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else if err == errs.ErrEditionNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-		}
-	case "version":
-		if err == errs.ErrDatasetNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else if err == errs.ErrEditionNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else if err == errs.ErrVersionNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-		}
-	case "dimension":
-		if err == errs.ErrDatasetNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else if err == errs.ErrEditionNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else if err == errs.ErrVersionNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else if err == errs.ErrDimensionsNotFound {
-			http.Error(w, err.Error(), http.StatusNotFound)
-		} else {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-		}
-	}
-}
-
-// PublishCheck Checks if an version has been published
-type PublishCheck struct {
-	Datastore store.Storer
-}
-
-// Check wraps a HTTP handle. Checks that the state is not published
-func (d *PublishCheck) Check(handle func(http.ResponseWriter, *http.Request)) http.HandlerFunc {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-		vars := mux.Vars(r)
-		id := vars["id"]
-		edition := vars["edition"]
-		version := vars["version"]
-
-		currentVersion, err := d.Datastore.GetVersion(id, edition, version, "")
-		if err != nil {
-			if err != errs.ErrVersionNotFound {
-				http.Error(w, err.Error(), http.StatusInternalServerError)
-				return
-			}
-			// If document cannot be found do not handle error
-			handle(w, r)
-			return
-		}
-
-		if currentVersion != nil {
-			if currentVersion.State == models.PublishedState {
-				defer func() {
-					if err := r.Body.Close(); err != nil {
-						log.ErrorC("could not close response body", err, nil)
-					}
-				}()
-
-				versionDoc, err := models.CreateVersion(r.Body)
-				if err != nil {
-					log.ErrorC("failed to model version resource based on request", err, log.Data{"dataset_id": id, "edition": edition, "version": version})
-					http.Error(w, err.Error(), http.StatusBadRequest)
-					return
-				}
-
-				// We can allow public download links to be modified by the exporters when a version is published.
-				// Note that a new version will be created which contain only the download information to prevent
-				// any forbidden fields from being set on the published version
-				if r.Method == "PUT" {
-					if versionDoc.Downloads != nil {
-						newVersion := new(models.Version)
-						if versionDoc.Downloads.CSV != nil && versionDoc.Downloads.CSV.Public != "" {
-							newVersion = &models.Version{
-								Downloads: &models.DownloadList{
-									CSV: &models.DownloadObject{
-										Public: versionDoc.Downloads.CSV.Public,
-										Size:   versionDoc.Downloads.CSV.Size,
-										HRef:   versionDoc.Downloads.CSV.HRef,
-									},
-								},
-							}
-						}
-						if versionDoc.Downloads.XLS != nil && versionDoc.Downloads.XLS.Public != "" {
-							newVersion = &models.Version{
-								Downloads: &models.DownloadList{
-									XLS: &models.DownloadObject{
-										Public: versionDoc.Downloads.XLS.Public,
-										Size:   versionDoc.Downloads.XLS.Size,
-										HRef:   versionDoc.Downloads.XLS.HRef,
-									},
-								},
-							}
-						}
-						if newVersion != nil {
-							b, err := json.Marshal(newVersion)
-							if err != nil {
-								http.Error(w, err.Error(), http.StatusForbidden)
-								return
-							}
-
-							if err := r.Body.Close(); err != nil {
-								log.ErrorC("could not close response body", err, nil)
-							}
-							r.Body = ioutil.NopCloser(bytes.NewBuffer(b))
-							handle(w, r)
-							return
-						}
-					}
-				}
-
-				err = errors.New("unable to update version as it has been published")
-				log.Error(err, log.Data{"version": currentVersion})
-				http.Error(w, err.Error(), http.StatusForbidden)
-				return
-			}
-		}
-
-		handle(w, r)
-	})
-}
-
-func setJSONContentType(w http.ResponseWriter) {
-	w.Header().Set("Content-Type", "application/json")
-}
-
-func (api *DatasetAPI) authenticate(r *http.Request, logData map[string]interface{}) (bool, map[string]interface{}) {
-	var authorised bool
-
-	if api.EnablePrePublishView {
-		var hasCallerIdentity, hasUserIdentity bool
-
-		callerIdentity := common.Caller(r.Context())
-		if callerIdentity != "" {
-			logData["caller_identity"] = callerIdentity
-			hasCallerIdentity = true
-		}
-
-		userIdentity := common.User(r.Context())
-		if userIdentity != "" {
-			logData["user_identity"] = userIdentity
-			hasUserIdentity = true
-		}
-
-		if hasCallerIdentity || hasUserIdentity {
-			authorised = true
-		}
-		logData["authenticated"] = authorised
-
-		return authorised, logData
-	}
-	return authorised, logData
-}
-
-func handleAuditingFailure(w http.ResponseWriter, err error, logData log.Data) {
-	log.ErrorC(auditError, err, logData)
-	http.Error(w, "internal server error", http.StatusInternalServerError)
 }
